feat: 주문에 쿠폰 할인 적용

+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductResult.java
@@ -21,7 +21,7 @@ public class ProductResult {
                     product.getName(),
                     product.getDescription(),
                     product.getImageUrl(),
-                    product.getPrice(),
+                    product.getPrice().longValue(),
                     product.getLikeCount()
             );
         }
