+++ apps/commerce-api/src/main/java/com/loopers/domain/order/OrderInfo.java
@@ -21,7 +21,7 @@ public class OrderInfo {
                     order.getId(),
                     order.getUserId(),
                     orderItemDtos,
-                    order.getTotalAmount().value(),
+                    order.getTotalAmount().value().longValue(),
                     order.getOrderDate(),
                     order.getStatus()
             );
@@ -35,7 +35,7 @@ public class OrderInfo {
             int quantity
     ) {
         public static OrderItemDto from(OrderItem orderItem) {
-            return new OrderItemDto(orderItem.getProductId(), orderItem.getName(), orderItem.getProductPrice().value(), orderItem.getQuantity());
+            return new OrderItemDto(orderItem.getProductId(), orderItem.getName(), orderItem.getProductPrice().value().longValue(), orderItem.getQuantity());
         }
     }
