+++ apps/commerce-api/src/main/java/com/loopers/domain/point/Point.java
@@ -9,6 +9,8 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "point")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -25,7 +27,7 @@ public class Point extends BaseEntity {
 
     protected Point(Long userId, Long balance) {
         this.userId = userId;
-        this.balance = new Money(balance);
+        this.balance = new Money(BigDecimal.valueOf(balance));
     }
 
     public static Point create(Long userId) {
@@ -39,14 +41,14 @@ public class Point extends BaseEntity {
         if(amount <= 0) {
             throw new CoreException(ErrorType.INVALID_POINT_AMOUNT, "충전할 포인트는 0원 이상이어야 합니다.");
         }
-        this.balance = this.balance.plus(amount);
+        this.balance = this.balance.plus(BigDecimal.valueOf(amount));
     }
 
     public void use(long amount) {
         if(amount <= 0) {
             throw new CoreException(ErrorType.INVALID_POINT_AMOUNT, "사용할 포인트는 0원 이상이어야 합니다.");
         }
-        this.balance = this.balance.minus(amount);
+        this.balance = this.balance.minus(BigDecimal.valueOf(amount));
     }
 
     public Money balance() {
